Poprawka typów argumentów funkcji display_ball()

+++ Core/Inc/acc_ball.h
@@ -14,14 +14,14 @@
 int pozycja_x=0;
 int pozycja_y=0;
 
-void display_ball(int accX, int accY){
+void display_ball(float accX, float accY){
 
 	int x = ILI9341_WIDTH * accX / 32768;
 	int y = ILI9341_HEIGHT * accY / 32768;
 
 	ILI9341_DrawFilledCircle(ILI9341_WIDTH / 2 + x, ILI9341_HEIGHT / 2 - y, 10, 90);
 }
-void acc_ball(int accX, int accY){
+void acc_ball(float accX, float accY){
 
 	int x = accX / 32768;
 	int y = accY / 32768;
